Move InviteAdminScreen to graphql/portal folder #498
